+++ JotunnLib/Managers/PieceManager.cs
@@ -562,12 +562,13 @@ namespace Jotunn.Managers
                 return;
             }
 
-            if (!PrefabManager.Instance.Prefabs.Any(x => x.Prefab.name.Equals(prefab.name)))
+            int hash = prefab.name.GetStableHashCode();
+
+            if (!PrefabManager.Instance.Prefabs.ContainsKey(hash))
             {
                 PrefabManager.Instance.AddPrefab(prefab);
             }
 
-            var hash = prefab.name.GetStableHashCode();
             if (ZNetScene.instance != null && !ZNetScene.instance.m_namedPrefabs.ContainsKey(hash))
             {
                 PrefabManager.Instance.RegisterToZNetScene(prefab);