+++ CHANGELOG.md
@@ -2,12 +2,14 @@
 
 ## Version 2.2.8
 * Added global mod registry, collecting added entities per mod
+* Added entity CustomPrefab including the possibility to let Jötunn fix mock references
 * Show color on the ColorPicker button in the mod settings
 * Refactored manually built controls in GUIManager to use Unity's DefaultControls (thx redseiko)
 * GUIManager.ApplyButtonStyle does not add a Text GO any more
 * Added GUIManager.CreateInputField
 * Fixed sprite atlas loading
 * Fixed OnLocalizationAdded event timing
+* Deprecated ItemManager.OnVanillaItemsAvailable, mods should use PrefabManager.OnVanillaObjectsAvailable now
 
 ## Version 2.2.7
 * Added loading of dll embedded text assets (thx MSchmoecker)